Modal upon deleting added, add product button added, reFetchData on create and delete operation added

+++ src/components/Table.vue
@@ -20,9 +20,9 @@
                                 </tr>
                             </thead>
                             <tbody>
-                                <tr
-                                    v-for="(product, index) in products" :key="product.id" class="transition duration-300 ease-in-out hover:bg-neutral-100 dark:border-neutral-500 dark:hover:bg-neutral-600 bg-[#FAFAFA]">
-                                    <td class="whitespace-nowrap px-6 py-4 font-medium">{{++index}}</td>
+                                <tr v-for="(product, index) in products" :key="product.id"
+                                    class="transition duration-300 ease-in-out hover:bg-neutral-100 dark:border-neutral-500 dark:hover:bg-neutral-600 bg-[#FAFAFA]">
+                                    <td class="whitespace-nowrap px-6 py-4 font-medium">{{+ + index }}</td>
                                     <td class="whitespace-nowrap px-6 py-4">{{ product.assetNumber }}</td>
                                     <td class="whitespace-nowrap px-6 py-4">{{ product.categoryName }}</td>
                                     <td class="whitespace-nowrap px-6 py-4">N/A</td>
@@ -33,8 +33,8 @@
                                     <td class="whitespace-nowrap px-6 py-4">{{ product.purchaseDate }}</td>
                                     <td class="whitespace-nowrap px-6 py-4">
                                         <div class="flex gap-2">
-                                            <PencilSquareIcon class="h-6 w-6 text-blue-500" @click="getRowData(product)"/>
-                                            <TrashIcon class="h-6 w-6 text-red-500" @click="deleteData(product.id)"/>
+                                            <PencilSquareIcon class="h-6 w-6 text-blue-500" @click="getRowData(product)" />
+                                            <TrashIcon class="h-6 w-6 text-red-500" @click="deleteData(product.id)" />
                                         </div>
                                     </td>
                                 </tr>
@@ -49,10 +49,21 @@
 
 <script setup>
 import axios from 'axios';
-import { ref } from 'vue';
+import { onUpdated, ref } from 'vue';
+import Swal from 'sweetalert2';
 import { PencilSquareIcon, TrashIcon } from "@heroicons/vue/24/outline";
 
+const props = defineProps({
+    reFetchData: Boolean
+});
+
 const emit = defineEmits(['editData']);
+
+onUpdated(() => {
+    if (props.reFetchData) {
+        getTableData();
+    }
+});
 
 const products = ref();
 
@@ -74,22 +85,33 @@ const getTableData = async () => {
 getTableData();
 
 const deleteData = async (id) => {
-    try {
-        const apiKey = 'IRyKCBuGQ1PpflCBs7ZaU+KImwTULz1fU8zjWE/aKhU=';
-        const url = `http://182.163.101.173:49029/product-crud/products/${id}`;
-        const response = await axios.delete(url, {
-            headers: {
-                'apiKey': apiKey
+    Swal.fire({
+        text: "Are you sure you want to delete this product?",
+        icon: 'error',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+    }).then( async (result) => {
+        if (result.isConfirmed) {
+            try {
+                const apiKey = 'IRyKCBuGQ1PpflCBs7ZaU+KImwTULz1fU8zjWE/aKhU=';
+                const url = `http://182.163.101.173:49029/product-crud/products/${id}`;
+                const response = await axios.delete(url, {
+                    headers: {
+                        'apiKey': apiKey
+                    }
+                });
+                getTableData();
+            } catch (error) {
+                console.error(error);
             }
-        });
-        getTableData();
-    } catch (error) {
-        console.error(error);
-    }
+        }
+    })
 };
 
 const getRowData = (rowData) => {
-    console.log(rowData.id);
+    console.log(rowData);
     emit('editData', rowData);
 }
